feat: componentizando listagem de editais
